backend/requirements.txt
django==5.2
djangorestframework==3.15.*
djangorestframework-simplejwt==5.3.*
django-cors-headers==4.3.*
psycopg2-binary==2.9.*
django-environ==0.11.*
drf-spectacular==0.27.*
django-filter==24.*
requests==2.31.*
Pillow==10.*

# Real-time (Django Channels + Redis)
channels==4.1.*
channels-redis==4.2.*
daphne==4.1.*

# Payments
razorpay==1.4.*
cashfree-pg==4.*

# PostGIS / geolocation
pgeocode==0.4.1
# NOTE: GDAL + GEOS are installed via system packages in the Dockerfile
# django.contrib.gis requires these system libs
